Move SubtitleListEditor to new file
